refactor(storage): 移除 webext-bridge 依赖，改为直接存储访问
重构了整体架构，简化了组件与服务层的交互方式：

- 移除 background script 中的消息处理器

- 组件直接调用服务层函数，提升性能

- 新增 queryAggBalanceRecords 聚合查询函数

- 整理 import 结构，提取公共工具函数

BREAKING CHANGE: 移除了 webext-bridge 消息传递机制
