+++ service/utils.ts
@@ -103,3 +103,14 @@ export const getHourStartTimestamp = (timestamp: number) => {
   date.setMinutes(0, 0, 0);
   return date.getTime();
 }
+
+// 格式化字节数为人类可读格式
+export const formatBytes = (bytes: number): string => {
+  if (bytes === 0) return '0 B';
+
+  const k = 1024;
+  const sizes = ['B', 'KB', 'MB', 'GB', 'TB'];
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+
+  return parseFloat((bytes / Math.pow(k, i)).toFixed(1)) + ' ' + sizes[i];
+}
